refactor: split blog composable into modular architecture
Refactored use-blog-enhanced.composable.ts (775 lines) into smaller,
more maintainable modules following single responsibility principle.

Changes:
- Created utility modules:
  * html-utils.ts: HTML cleaning and entity decoding
  * text-utils.ts: Text manipulation and reading time estimation
  * date-utils.ts: Date formatting utilities
  * cache-utils.ts: localStorage cache management

- Created specialized composables:
  * use-blog-core.composable.ts: Core blog state and loading logic
  * use-blog-filters.composable.ts: Search, filtering, and pagination
  * use-blog-categories.composable.ts: Category management
  * use-blog-features.composable.ts: Favorites and sharing features

- Refactored main composable:
  * use-blog-enhanced.composable.ts: Now orchestrates specialized modules
  * Maintains same public API for backward compatibility
  * Reduced from 775 lines to ~125 lines

Benefits:
✅ Better code organization and maintainability
✅ Easier to test individual modules
✅ Eliminated duplicate code (decodeHtmlEntities was defined twice)
✅ Improved code reusability
✅ Better separation of concerns

Build: ✅ Successful (warnings only, no errors)
